+++ pypeit/par/pypeitpar.py
@@ -1281,7 +1281,7 @@ class WavelengthSolutionPar(ParSet):
         """
         Return the valid wavelength solution methods.
         """
-        return [ 'simple', 'semi-brute', 'basic', 'holy-grail', 'reidentify', 'full_template']
+        return [ 'simple', 'semi-brute', 'basic', 'holy-grail', 'identify', 'reidentify', 'full_template']
 
     @staticmethod
     def valid_lamps():
@@ -1486,7 +1486,7 @@ class TraceSlitsPar(ParSet):
     def from_dict(cls, cfg):
         k = cfg.keys()
         parkeys = ['function', 'medrep', 'number', 'trim', 'maxgap', 'maxshift', 'pad',
-                   'sigdetect', 'min_slit_width', 'diffpolyorder', 'sobel_mode', 
+                   'sigdetect', 'min_slit_width', 'diffpolyorder', 'sobel_mode',
                    'pcaextrap', 'add_slits', 'rm_slits', 'smash_range', 'trace_npoly',
                    'mask_frac_thresh']
         kwargs = {}
@@ -2942,7 +2942,7 @@ class DetectorPar(ParSet):
         # TODO: Allow for None, such that there is no overscan region
         defaults['oscansec'] = 'BIASSEC' if pars['numamplifiers'] is None \
                                         else ['BIASSEC']*pars['numamplifiers']
-        dtypes['oscansec'] = [str, list]
+        dtypes['oscansec'] = [str, list, type(None)]
         descr['oscansec'] = 'Either the overscan section or the header keyword where the valid ' \
                             'data sections can be obtained, one per amplifier. If defined ' \
                             'explicitly should be in FITS format (e.g., [1:2048,10:4096]).'
@@ -2978,7 +2978,7 @@ class DetectorPar(ParSet):
         """
         if self.data['numamplifiers'] > 1:
             keys = [ 'gain', 'ronoise', 'datasec', 'oscansec' ]
-            dtype = [ (int, float), (int, float), str, str ]
+            dtype = [ (int, float), (int, float), str, (str, None) ]
             for i in range(len(keys)):
                 if self.data[keys[i]] is None:
                     continue